Code review and fixes
