+++ document-legacy/src/document_metadata.rs
@@ -1,13 +1,12 @@
-use glam::{DAffine2, DVec2};
+use graph_craft::document::{DocumentNode, NodeId, NodeNetwork};
 use graphene_core::renderer::ClickTarget;
+use graphene_core::renderer::Quad;
 use graphene_core::transform::Footprint;
 use graphene_core::uuid::ManipulatorGroupId;
+
+use glam::{DAffine2, DVec2};
 use std::collections::{HashMap, HashSet};
 use std::num::NonZeroU64;
-
-use graph_craft::document::{DocumentNode, NodeId, NodeNetwork};
-
-use graphene_core::renderer::Quad;
 
 #[derive(Debug, Clone)]
 pub struct DocumentMetadata {
