+++ editor/src/messages/portfolio/document/overlays/mod.rs
@@ -1,5 +1,7 @@
 mod overlays_message;
 mod overlays_message_handler;
+pub mod utility_functions;
+pub mod utility_types;
 
 #[doc(inline)]
 pub use overlays_message::*;
