+++ editor/src/messages/portfolio/document/overlays/overlays_message.rs
@@ -1,19 +1,14 @@
-use crate::messages::portfolio::document::overlays::overlays_message_handler::OverlayContext;
+use super::utility_types::{empty_provider, OverlayProvider};
 use crate::messages::prelude::*;
 
 use serde::{Deserialize, Serialize};
 
-pub type OverlayProvider = fn(OverlayContext) -> Message;
-fn empty_provider() -> OverlayProvider {
-	|_| Message::NoOp
-}
-
 #[impl_message(Message, DocumentMessage, Overlays)]
 #[derive(PartialEq, Clone, Debug, Serialize, Deserialize)]
 pub enum OverlaysMessage {
-	Render,
+	Draw,
 
-	// I don't know why we need to serde messages - we never use this functionality
+	// Serde functionality isn't used but is required by the message system macros
 	AddProvider(#[serde(skip, default = "empty_provider")] OverlayProvider),
 	RemoveProvider(#[serde(skip, default = "empty_provider")] OverlayProvider),
 }
