+++ editor/src/messages/tool/common_functionality/mod.rs
@@ -1,6 +1,5 @@
 pub mod color_selector;
 pub mod graph_modification_utils;
-pub mod overlay_renderer;
 pub mod pivot;
 pub mod resize;
 pub mod shape_editor;
